requirements.txt
pytest==6.2.5
selenium==3.141.0
pytest-xdist==2.1.0
allure-pytest==2.8.22
requests==2.24.0
mysql-connector-python==8.0.22
webdriver-manager==3.7.1
Faker==14.2.1
python-dotenv==0.19.2
